uniswap-v2-swap-bot/requirements.txt
web3==6.20.2
python-dotenv==1.0.1
